requirements.txt
Django
python-telegram-bot
requests
